+++ karbide-core/src/commonMain/kotlin/dev/karmakrafts/karbide/BitOrder.kt
@@ -1,0 +1,5 @@
+package dev.karmakrafts.karbide
+
+enum class BitOrder {
+    MSB_FIRST, LSB_FIRST
+}